# api_ui/api.yaml
openapi: 3.1.0
info:
  title: Polygon ID - Issuer - UI API
  description: |
    Documentation for the Issuer - UI API
  version: "1"
servers:
  - description: Local
    url: http://localhost:3002
...
paths:
...
  /v1/credentials/revocation/status/{nonce}:
    get:
      summary: Get Revocation Status
      operationId: GetRevocationStatus
      description: Endpoint to get the revocation status of a credential
      tags:
        - Credential
      parameters:
        - $ref: '#/components/parameters/pathNonce'
      responses:
        '200':
          description: Proof
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevocationStatusResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
  /v1/qr-store:
    get:
      summary: QrCode body
      operationId: GetQrFromStore
      description: Returns a previously generated QR code via url shortener method
      tags:
        - Agent
      parameters:
        - in: query
          name: id
          schema:
            type: string
            x-go-type: uuid.UUID
            x-go-type-import:
              name: uuid
              path: github.com/google/uuid
            example: 8edd8112-c415-11ed-b036-debe37e1cbd6
      responses:
        '200':
          description: A json to generate a QR code
          content:
            application/json:
              schema:
                type: object
        '400':
          $ref: '#/components/responses/400'
        '404':
          $ref: '#/components/responses/404'
        '500':
          $ref: '#/components/responses/500'

  #state:
  /v1/state/publish:
    post:
      summary: Publish Identity State
      operationId: PublishState
      security:
        - basicAuth: [ ]
      tags:
        - State
      responses:
        '202':
          description: Publish state successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishIdentityStateResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
  /v1/state/retry:
    post:
      summary: Retry Publish Identity State
      operationId: RetryPublishState
      security:
        - basicAuth: [ ]
      tags:
        - State
      responses:
        '202':
          description: Publish state successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishIdentityStateResponse'
        '400':
          $ref: '#/components/responses/400'
        '500':
          $ref: '#/components/responses/500'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
...
    GenericErrorMessage:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: 'Something happen'
...
    PublishIdentityStateResponse:
      type: object
      properties:
        txID:
          type: string
        claimsTreeRoot:
          type: string
        state:
          type: string
        revocationTreeRoot:
          type: string
        rootOfRoots:
          type: string
      example:
        claimsTreeRoot: "3608a07d1d455974054065d6138bb4e634afa7a965f29919f4d0f494eba0f91c"
        revocationTreeRoot: "0000000000000000000000000000000000000000000000000000000000000000"
        rootOfRoots: "53b5e088b1c93b2ea4748bc29649c1cd6b0f4749bab3f96d2d9f0aac8eab1b19"
        state: "8d0dfb1b7bc910e347efbba324e604359815c40b56b77e191fdac1eb7f770119"
        txID: "0x45aef0730854606bf9ea3cabba80541fa3dc61833c7a08b6c722d732451fea46"
...
    RevocationStatusResponse:
      type: object
      required:
        - issuer
        - mtp
      properties:
        issuer:
          type: object
          properties:
            state:
              type: string
            rootOfRoots:
              type: string
            claimsTreeRoot:
              type: string
            revocationTreeRoot:
              type: string
        mtp:
          type: object
          required:
            - existence
          properties:
            existence:
              type: boolean
            siblings:
              type: array
              x-omitempty: false
              items:
                type: string

            node_aux:
              type: object
              properties:
                key:
                  type: string
                value:
                  type: string
...
  parameters:
...
    pathNonce:
      name: nonce
      in: path
      required: true
      description: |
        Claim nonce, e.g: 4182463773
      schema:
        type: integer
        format: int64
  responses:
    '400':
      description: 'Bad Request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'
...
    '404':
      description: 'Entity not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'
...
    '500':
      description: 'Internal Server error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericErrorMessage'